requirements.txt
# Flask Framework
Flask==3.0.0
Flask-Cors==4.0.0

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Data Processing
python-dateutil==2.8.2

# Caching
Flask-Caching==2.1.0

# Rate Limiting
Flask-Limiter==3.5.0

# JSON Processing
jsonschema==4.20.0

# IP Address Validation
ipaddress==1.0.23

# Async Support (optional)
aiohttp==3.9.1

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2

# PDF Generation
reportlab==4.4.4
pillow==12.0.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Production Server
gunicorn==21.2.0

# Utilities
colorama==0.4.6
tabulate==0.9.0
